+++ Packages/webxr-interactions/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- No reference to controllers in SceneHitTest.
+
 ## [0.16.0] - 2023-02-02
 ### Added
 - Spectator Camera.